pyproject.toml
[project]
name = "metakeyai-server"
version = "0.1.0"
description = "MetaKeyAI Python server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "dspy-ai>=2.4.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "nltk>=3.8.1",
    "python-dateutil>=2.8.2",
    "regex>=2023.0.0",
    "langdetect>=1.0.9"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/python_scripts"]

[tool.uv]
dev-dependencies = []
